[install] sweetalert2

+++ resources/js/pages/Jobs.vue
@@ -11,7 +11,9 @@
         v-slot="{ rows }"
       >
         <base-table-row v-for="row in rows" :key="row.id">
-          <base-table-column>{{ formatDate(row.created_at) }}</base-table-column>
+          <base-table-column>{{
+            formatDate(row.created_at)
+          }}</base-table-column>
           <base-table-column>{{ row.name }}</base-table-column>
           <base-table-column class="text-end" overflow-visible>
             <action-tool
@@ -25,7 +27,7 @@
     </Box>
 
     <job-set-data-modal
-      ref="jobModal" 
+      ref="jobModal"
       :value.sync="showModalJob"
       @action-done="$refs.table.fetchs()"
     />
@@ -33,7 +35,7 @@
 </template>
 
 <script>
-import JobSetDataModal from '@components/modal/JobSetDataModal.vue'
+import JobSetDataModal from "@components/modal/JobSetDataModal.vue";
 
 export default {
   components: { JobSetDataModal },
@@ -48,18 +50,50 @@ export default {
         name: "name",
         class: "",
       },
-    ]
+    ],
   }),
   methods: {
     edit(id) {
-      this.$refs.jobModal.editId = id
-      this.showModalJob = true
+      this.$refs.jobModal.editId = id;
+      this.showModalJob = true;
     },
     async remove(id) {
-      await axios.delete(`/api/job/${id}`, this.$root.headersRequestAPI)
+      const swalWithBootstrapButtons = this.$swal.mixin({
+        customClass: {
+          confirmButton: "btn btn-success",
+          cancelButton: "btn btn-danger",
+        },
+        buttonsStyling: false,
+      });
 
-      this.$refs.table.fetchs()
-    }
-  }
-}
+      this.$swal({
+        title: "Are you sure?",
+        text: "will you remove this?",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonText: "Yes, delete it!",
+        cancelButtonText: "No, cancel!",
+        reverseButtons: true,
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          await axios.delete(`/api/job/${id}`, this.$root.headersRequestAPI);
+
+          this.$refs.table.fetchs();
+
+          swalWithBootstrapButtons.fire(
+            "Deleted!",
+            "success deleted.",
+            "success"
+          );
+        } else if (result.dismiss === this.$swal.DismissReason.cancel) {
+          swalWithBootstrapButtons.fire(
+            "Cancelled",
+            "Your data is safe :)",
+            "error"
+          );
+        }
+      });
+    },
+  },
+};
 </script>
